+++ src/App.css
@@ -1,7 +1,6 @@
 @import url('https://fonts.googleapis.com/css?family=Josefin+Sans');
 
 .App {
-  /*text-align: center;*/
   background-color: #00b2e4;
 }
 
@@ -16,7 +15,6 @@
 }
 
 .App-logo {
-  /*height: 40vmin;*/
   position: absolute;
   left: 70px;
   top: 30px;
@@ -40,65 +38,4 @@ h1{
   position: absolute;
   top: 30px;
   left: 590px;
-}
-
-
-.dropdown-all{
-  color: #0c4164;
-}
-
-.dropdown-assessmentType{
-  position: absolute;
-  top:230px;
-  left:150px;
-}
-
-.dropdown-race{
-  position: absolute;
-  top: 230px;
-  left:800px
-}
-
-.dropdown-gender{
-  position: absolute;
-  top: 230px;
-  left: 500px
-}
-
-.dropdown-typeOfLoss{
-  position: absolute;
-  top: 315px;
-  left: 150px;
-}
-
-.dropdown-level{
-  position: absolute;
-  top:315px;
-  left: 500px
-}
-
-.dropdown-prePost{
-  position: absolute;
-  top:315px;
-  left: 800px;
-}
-
-button{
-  font-family: sans-serif;
-  background-color: #f7df37;
-  color: #ff3d00;
-  padding: 10px 60px;
-  position: absolute;
-  right: 150px;
-  top: 315px;
-  font-size: 20px;
-  font-weight: bold;
-  border-color: #0c4164;
-  border-radius: 4px;
-  border-width: 1px;
-}
-
-/* shadow button code: https://www.w3schools.com/css/css3_buttons.asp */
-.buttonShadow:hover {
-  box-shadow: 0 12px 16px 0 rgba(0,0,0,0.24),0 17px 50px 0 rgba(0,0,0,0.19);
 }
